Telegram-Instaloader
This program runs bot in telegram, wich can parse people's profiles. As a parsing module i use instaloader. If you want to download all pictures, videos, subscribes and more, you need to run main file, so it could run up my bot. Also you need to enter to your profile and be subscribed to people you want to download, if there profile is closed( So this way you can't parse closed profile if you not subscribed on it). Use it and share with your friends and get enjoy!